Replaced data_object types "TSCE" and "TSCG" by "TS" and added optional "do_cg" field.

+++ include/hnzconf.h
@@ -129,7 +129,7 @@ class HNZConf {
   /**
    * Get the number of CG. Used for the consistency check when GI.
    */
-  int getNumberCG() const;
+  unsigned long getNumberCG() const;
 
   /**
    * Get the IP address to remote IEC 104 server (A path)
